Rename src/{rooms.tsx => compat.ts}
