+++ packages/liveblocks-react/src/compat.ts
